Add `ProblemJsonResponse` wrapper around `Response`
This allows us to pass in a `Response` object (created from exception handling),
and get back out a `Response` object which has been formatted to problem+json
standards.

This is very much a WIP/POC and we should standardize what we want for error response
payloads and map them to different errors and exceptions we'll raise, or expect
to be raised from DRF.

This does add a default as well.