Add a WorkspaceValidationService object
This adds a `WorkspaceValidationService` which can be used to offload the logic
for doing `Workspace` validations either from the view, serializer, model, or elsewhere
without being tied to the caller's context.

One potential smell is that since we'll use this in the `Workspace` model, we'd
have a circular dep unless we load the model this way.

This should be easily testable, and can be used to consolidate and simplify the
view, serializer and model logic.