Optimize query for retrieving descendant IDs based on resource definitions
Instead of having to iterate over each permission, and each workspace related to
the resource definition IDs in that permission, generating 1 to many (potentially
redundant) queries, this change generalizes the CTE query we already had for descendants
of a single workspace instance.

Now, this query accepts an array, and will recursively collect all _distinct_
values for all workspaces found in the ID list, _including_ the roots.

This means we can get back the list of IDs in a single query regardless of the
number of resource definition IDs. We can also reuse this in the `descendants`
instance/model method by passing in only the instance ID, and excluding the
instance from the resultant list before filtering to return descendants. This will
however introduce an additional query for `descendants()`, so it may be worth
leaving it as its own CTE vs trying to DRY things up.