Only return hierarchy on `/access/` requests, ensuring it's in a list
Since the `ResourceDefinitionSerializer` is reused by both the `AccessSerializer` directly,
and `RoleSerializer` (since it uses `AccessSerializer`), we want to only ever include
hierarchy in access when the `/access/` endpoint is called, not on `/roles/`.

This adds metadata to the `context` objects in the `AccessView` to ensure we
know when the `AccessSerializer` is being instantiated from the `AccessView` vs
`RoleViewSet`. Then in the `ResourceDefinitionSerializer` we're able to add that
to our contitional check.

This also explicitly ensures the resource definition being returned has an `in`
operation value when including hierarchy, since we'll always convert the `value`
to a list.