Optimize hierarchy serializer queries
With the conditional inclusion of hierarchy, this introduced additional queries
in the serializer when we'd check the accesses permission object for each resource
definition on a role (creating an n+1 problem) to check the application and resource
type to know whether or not we should include hierarchy in the response.

This change updates the access queryset so that we select the permission of the
access objects in the query, as well as prefect the resource definitions, ensuring
we select the access permissions related. This effectively removed all additional
(n) queries on resource definitions to get permission objects.