Add a V2ResultsSetPagination class to our paginator module
This adds a `V2ResultsSetPagination` class, which inherits from `StandardResultsSetPagination`
and overrides the `paginate_queryset` method.

We do this to check the `?limit=` query param on a paginated API request and check
it against `NO_LIMIT_ENFORCED_VALUE` which is `-1`. If the limit value passed in
matches, we'll do two things:

- unset the `max_limit` on the paginator so that we never hit the default upper bound.
- set the `default_limit` on the paginator equal to the length of the queryset, so that
  the pagination limit accurately reflects the number of records.

We then set `pagination_class = V2ResultsSetPagination` in the base v2 viewset,
so that v2 APIs (including Workspaces) include this functionality by default.