feature: read UMB certificates from the "automatic renewal" secret
In order to avoid breaking the existing set up, a temporary flag has
been introduced which will allow us to switch on and off reading from
the secret that will keep the automatically renewed certificates.

Since the secret contains the certificate and the CA bundles separated,
we need to produce a third file that combines the two whenever we use
the automatically renewed certificates.

RHCLOUD-40635