Ensure `username` is set on system user generated from JWT
In the middleware, in order to set the `request.user = user` we first check to ensure
there's a `username` and `user.org_id` on the `user` object [1].

This currently causes JWT system users to be created without a `username` since
it's based on `preferred_username` [2] which does not exist on the client service
account tokens. This ultimately means `request.user` is not the custom user object
we've created, but an anonymous user [3] which does not have the fields required
to fulfill requests [4] resulting in 500s.

This adds the client/user id value as the username in the builder logic.

[1] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/rbac/middleware.py#L304-L305
[2] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/management/authorization/token_validator.py#L63
[3] https://docs.djangoproject.com/en/4.2/ref/contrib/auth/
[4] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/management/permissions/workspace_access.py#L28