Fix order of operations when paginating data
Previously we were serializing the data first, and then paginating that serialized
data. This is an antipattern, and caused an issue when making the update to use
`queryset.count()` vs `len(queryset)` since it was a list, not a queryset.

This ensures we first paginate the queryset, then serialize the paginated data.