Remove workspace relation in get_kessel_relation_tuples
The parent-child relationship between workspaces does not belong to the
role, so it is improper to add it in get_kessel_relation_tuples. This
would be harmless if we were just adding it (since it is supposed to
exist, and, at worst, we would be attempting to re-create it when it
already exists). However, in RelationApiDualWriteHandler, we also use
this function (through migrate_role) to generate the set of relations to
remove when updating or removing a role. So, if we call
replicate_deleted_role, we will also remove the workspace-workspace
relation when attempting to remove the role; this would cause the
relations in Kessel to be out-of-sync with the RBAC database.