Use get_kessel_relation_tuples in RelationApiDualWriteHandler.prepare_for_update
The migrate_role function explicitly performs non-deterministic
operations and mutates its inputs. This is unlikely to be what we want
when calculating what relations to remove (because we always need to
calculate exactly the relations that we added).

permission_groupings_to_v2_role_bindings does appear to maintain a
consistent state when run on its own previous output (but would not if,
e.g., two v2 roles corresponding to the same v1 role ended up having the
same set of permissions, which is representable in the database, but is
not a possible output of migrate_role). It maintains the invariant that
only one v2 role will be created for any given set of permissions.
Thus, for each encountered role R, it will either find a cached role
with the same permissions in latest_roles_by_id (which, because of the
invariant, must be R itself), or (if R has not been seen before) it will
reuse R's id and update R to have the same permissions it already has
(because the v1 permissions that
permission_groupings_to_v2_role_bindings reads will not have changed in
the intervening time, since prepare_for_update is called before every
update).

Similarly, since the set of BindingMappings will not have changed since
the last time the v1 role was synced, the current BindingMapping will
always be reused (and not changed).

Thus, calling migrate_role in prepare_for_update for update is exactly
the same as calling get_kessel_relation_tuples on the current set of
role bindings. Changing prepare_for_update to directly call
get_kessel_relation_tuples much more clearly indicates this intent: the
definition of get_kessel_relation_tuples is obviously deterministic (and
thus does not carry any risk of resulting in the database getting out of
sync), and the purpose of prepare_for_update is to determine what
relations were added in the past, not to bring the models into a new
state (as the word "migrate" suggests).