feature: cache principals in the "/access" endpoint
We are having performance issues because every time the "/access"
endpoint gets called, the principal is fetched from the database, when
in reality those principals almost never change.

These changes store the principal that called the "/access" endpoint in
cache so that subsequent calls are faster, and we can spare some
database load.

RHCLOUD-40653