Actually create principals before cross-access request in test_dual_write
It appears to me that, in the real system, we will never create a cross-access request for a principal that does not exist. (This assumption *may* be wrong, in which case this change is also wrong, but surely we do not want to grant access to a user that we have no record of? Since the user_id in a cross-account request is the creating user, we would have to have an internal user authorized to create cross-account requests but that RBAC has no record of or that has since been deleted. In this case, it seems like it would be *good* to prevent approving the cross-account request?)

So, I believe the previous tests were doing something out-of-contract that happened to work due to the implementation of BindingMapping.

Once we have added a Principal to a RoleBinding, we have to remove that before we delete the Principal. This cannot occur in principal_removal (internal/views.py) because that will not delete any user with a user_id. It could occur in clean_tenant_principals (management/principal/cleaner.py), but I have confirmed with JayZ that it would be acceptable to add cleanup there.