refactor: allow specifying "principal_type" query parameter
Until now when querying for groups' service accounts the only allowed
query parameters were the "client ID", the "limit" and the "offset".
Even though when using the "client ID" query parameter it is assumed
that the desired principal types to be fetched are service accounts, the
UI also started sending the "principal type" query parameter, which
technically is not incorrect to send to the endpoint.

These changes should prevent the endpoint from returning a "400 - Bad
request" response when the "principal type" query parameter is specified
with a "service account" value.

RHCLOUD-41605