Ensure that removing a custom default access group always fully restores the default role binding
Currently, the following sequence events results in Relations being incorrect:
* A tenant creates a custom default access group.
* The V2 migration happens, where we bootstrap all existing tenants. (This has already happened.)
* The tenant deletes the custom default access group.

After this, the relationship from the default workspace to the default role binding will exist in Relations, but the relationships from that role binding to the platform default role and platform default group will no longer exist.

In order to fix this, we can always add all of the default role binding relationships when removing the custom default group. (This won't cause any issues because repeatedly adding a relationship is idempotent.)

Note that this cannot occur when bootstrapping new tenants, as they will not have a custom default access group at the time of bootstrap, and the default role binding will always be created.