RHCLOUD-42317: Always check for existing default groups when bootstrapping tenants (#1979)
Currently, if we bootstrap a single tenant, we do not check for an existing platform default group. In production, this isn't a problem, since the only time we bootstrap single tenants is when we have a brand new tenant (which necessarily does not have a platform default group).

Unfortunately, this ended up interfering with a test attempting to bootstrap a single tenant with a pre-existing custom default group. So, instead, always check for a platform default group.

This effectively reverts commit 2cea390a163485f1b4188bbd248fe4ed5a22de51.