RHCLOUD-41068: Consider scope when binding roles to groups (#2043)
When creating role bindings for groups, bind each role to the appropriate scope (as determined by the role's permissions).

Also remove unused generate_relations_to_add_roles in RelationApiDualWriteGroupHandler, since it complicates the implementation of this change.

Note that this makes no attempt to update existing bindings; that will be handled later.